+++ src/lib/components/MobileGuard.tsx
@@ -28,7 +28,8 @@ export const MobileGuard = ({ children }: MobileGuardProps) => {
     pathName.includes(`/contracts`) ||
     pathName.includes(`/codes`) ||
     // move
-    pathName.includes(`/modules`);
+    pathName.includes(`/modules`) ||
+    pathName.includes(`/validator`);
 
   if (!isResponsive && isMobile) return <NoMobile />;
   return <>{children}</>;