+++ src/lib/components/Tooltip.tsx
@@ -1,5 +1,5 @@
 import type { TooltipProps } from "@chakra-ui/react";
-import { Flex, Tooltip as ChakraTooltip } from "@chakra-ui/react";
+import { Tooltip as ChakraTooltip } from "@chakra-ui/react";
 
 import { CustomIcon } from "./icon";
 
@@ -11,6 +11,7 @@ export const Tooltip = ({
     hasArrow
     placement={placement}
     arrowSize={8}
+    shouldWrapChildren
     {...tooltipProps}
   />
 );
@@ -23,14 +24,13 @@ export const TooltipInfo = ({
   iconVariant = "default",
   ...tooltipProps
 }: TooltipInfoProps) => (
-  <Tooltip {...tooltipProps}>
-    <Flex cursor="pointer">
-      <CustomIcon
-        color="gray.600"
-        name={iconVariant === "solid" ? "info-circle-solid" : "info-circle"}
-        boxSize={3}
-        m={0}
-      />
-    </Flex>
+  <Tooltip {...tooltipProps} shouldWrapChildren>
+    <CustomIcon
+      color="gray.600"
+      name={iconVariant === "solid" ? "info-circle-solid" : "info-circle"}
+      boxSize={3}
+      m={0}
+      cursor="pointer"
+    />
   </Tooltip>
 );