+++ src/lib/pages/proposal-details/index.tsx
@@ -2,20 +2,18 @@ import { TabList, TabPanel, TabPanels, Tabs } from "@chakra-ui/react";
 import { useRouter } from "next/router";
 import { useCallback, useEffect } from "react";
 
+import { AmpEvent, track } from "lib/amplitude";
 import { useGovConfig, useInternalNavigate } from "lib/app-provider";
 import { CustomTab } from "lib/components/CustomTab";
 import { Loading } from "lib/components/Loading";
 import PageContainer from "lib/components/PageContainer";
 import { ErrorFetching, InvalidState } from "lib/components/state";
-import {
-  useProposalParams,
-  useProposalVotesInfo,
-} from "lib/services/proposalService";
+import { useProposalVotesInfo } from "lib/services/proposalService";
 
-import { ProposalTop, ProposalOverview, VoteDetail } from "./components";
-import { useDerivedProposalData } from "./data";
-import type { ProposalDetailsQueryParams } from "./type";
-import { zProposalDetailsQueryParams, TabIndex } from "./type";
+import { ProposalTop, ProposalOverview, VoteDetails } from "./components";
+import { useDerivedProposalData, useDerivedProposalParams } from "./data";
+import type { ProposalDetailsQueryParams } from "./types";
+import { zProposalDetailsQueryParams, TabIndex } from "./types";
 
 const ProposalDetailsBody = ({
   id,
@@ -24,12 +22,12 @@ const ProposalDetailsBody = ({
 }: ProposalDetailsQueryParams) => {
   useGovConfig({ shouldRedirect: true });
 
-  const router = useRouter();
   const navigate = useInternalNavigate();
   const { data, isLoading } = useDerivedProposalData(id);
   const { data: votesInfo, isLoading: isVotesInfoLoading } =
     useProposalVotesInfo(id);
-  const { data: params, isLoading: isParamsLoading } = useProposalParams();
+  const { data: params, isLoading: isParamsLoading } =
+    useDerivedProposalParams();
 
   const handleTabChange = useCallback(
     (nextTab: TabIndex) => () => {
@@ -47,22 +45,6 @@ const ProposalDetailsBody = ({
     },
     [id, tab, navigate]
   );
-
-  useEffect(() => {
-    if (router.isReady && (!tab || !Object.values(TabIndex).includes(tab))) {
-      navigate({
-        replace: true,
-        pathname: "/proposals/[id]/[tab]",
-        query: {
-          id,
-          tab: TabIndex.Overview,
-        },
-        options: {
-          shallow: true,
-        },
-      });
-    }
-  }, [router.isReady, tab, navigate, id]);
 
   if (isLoading) return <Loading />;
   if (!data) return <ErrorFetching dataName="proposal information" />;
@@ -85,7 +67,7 @@ const ProposalDetailsBody = ({
             Proposal Overview
           </CustomTab>
           <CustomTab onClick={handleTabChange(TabIndex.Vote)}>
-            Vote Detail
+            Vote Details
           </CustomTab>
         </TabList>
         <TabPanels>
@@ -98,7 +80,12 @@ const ProposalDetailsBody = ({
             />
           </TabPanel>
           <TabPanel p={0}>
-            <VoteDetail />
+            <VoteDetails
+              proposalData={data.info}
+              votesInfo={votesInfo}
+              params={params}
+              isLoading={isVotesInfoLoading || isParamsLoading}
+            />
           </TabPanel>
         </TabPanels>
       </Tabs>
@@ -110,6 +97,12 @@ const ProposalDetails = () => {
   const router = useRouter();
 
   const validated = zProposalDetailsQueryParams.safeParse(router.query);
+
+  useEffect(() => {
+    if (router.isReady && validated.success)
+      track(AmpEvent.TO_PROPOSAL_DETAILS, { tab: validated.data.tab });
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [router.isReady]);
 
   return (
     <PageContainer>